+++ lib/src/sync/sync_controller.dart
@@ -10,30 +10,74 @@ class SyncController {
     required ObjectBoxHelper objectBox,
     required String token,
     required RootIsolateToken rootToken,
+    String? primaryKey,
+    String? endPoint,
   }) async {
     final box = objectBox.outboxBox;
+    final refreshToken = objectBox.refreshToken;
+    final refreshTokenUrl = objectBox.refreshTokenUrl;
+    final clientId = objectBox.clientId;
+    final clientSecret = objectBox.clientSecret;
 
-    final items = box.query(OutboxItem_.isSynced.equals(0))
-        .order(OutboxItem_.priority , flags: Order.descending)
-        .order(OutboxItem_.createdAt)
-        .build()
-        .find();
+    List<OutboxItem> items = [];
+    if(primaryKey==null && endPoint==null) {
+      items = box.query(OutboxItem_.isSynced.equals(0))
+          .order(OutboxItem_.priority, flags: Order.descending)
+          .order(OutboxItem_.createdAt)
+          .build()
+          .find();
+    }
+    else if(primaryKey!=null && endPoint==null) {
+      items = box.query(
+          OutboxItem_.isSynced.equals(0).and(
+              OutboxItem_.primaryKey.equals(primaryKey)
+          )
+      )
+          .order(OutboxItem_.priority, flags: Order.descending)
+          .order(OutboxItem_.createdAt)
+          .build()
+          .find();
+
+    }
+    else if(primaryKey!=null && endPoint!=null) {
+      items = box.query(
+          OutboxItem_.isSynced.equals(0).and(
+              OutboxItem_.primaryKey.equals(primaryKey)
+          ).and(
+              OutboxItem_.endPoint.equals(endPoint)
+          )
+      )
+          .order(OutboxItem_.priority, flags: Order.descending)
+          .order(OutboxItem_.createdAt)
+          .build()
+          .find();
+    }
+
 
     final receivePort = ReceivePort();
 
     Isolate.spawn(
       outboxIsolateEntry,
-      OutboxIsolatePayload(items, token, rootToken, receivePort.sendPort),
+      OutboxIsolatePayload(items, token, rootToken, receivePort.sendPort,refreshToken,refreshTokenUrl,clientId,clientSecret),
     );
 
     receivePort.listen((msg) {
       if (msg is OutboxIsolateResult) {
         final item = box.get(msg.id);
         if (item != null) {
-          if (msg.response != null) {
-            item.response = jsonEncode(msg.response);
+
+          // 🟢 Save new token if received
+          if (msg.newAccessToken != null) {
+            item.newAccessToken = msg.newAccessToken;
+          }
+
+          item.statusCode = msg.statusCode;
+
+          if (msg.statusCode >= 200 && msg.statusCode < 300) {
+            item.response = msg.response;
             item.isSynced = 1;
           }
+
           item.retryCount++;
           item.lastTried = DateTime.now().millisecondsSinceEpoch;
           box.put(item);